Incorporate Lifecycle-Aware Components 1. Before you begin Architecture components are a set of Android libraries that help you structure your app in a way that is robust, testable, and maintainable.
This codelab introduces you to the following lifecycle-aware architecture components for building Android apps:

ViewModel - provides a way to create and retrieve objects that are bound to a specific lifecycle. A ViewModel typically stores the state of a view's data and communicates with other components, such as data repositories or the domain layer which handles business logic.
To read an introductory guide to this topic, see ViewModel.
LifecycleOwner - LifecycleOwner is an interface implemented by the AppCompatActivity and Fragment classes. You can subscribe other components to owner objects which implement this interface, to observe changes to the lifecycle of the owner.
To read an introductory guide to this topic, see Handling Lifecycles.
LiveData - allows you to observe changes to data across multiple components of your app without creating explicit, rigid dependency paths between them. LiveData respects the complex lifecycles of your app components, including activities, fragments, services, or any LifecycleOwner defined in your app. LiveData manages observer subscriptions by pausing subscriptions to stopped LifecycleOwner objects, and cancelling subscriptions to LifecycleOwner objects that are finished.
To read an introductory guide to this topic, see LiveData.
Prerequisites
This codelab has been designed for Android developers with some basic experience. To proceed, you must be familiar with the Android activity lifecycle.

What you'll build
In this codelab, you implement examples of each of the components described above. You begin with a sample app and add code through a series of steps, integrating the various architecture components as you progress.

What you'll need
Android Studio 3.5 or greater.